+++ src/manifold.h
@@ -12,14 +12,14 @@ class Manifold
 {
   std::unique_ptr<double[]> _flat = nullptr;
   std::vector<double> _y;
-  size_t _nobs, _E_x, _E_dt, _E_extras, _E_actual;
+  int _nobs, _E_x, _E_dt, _E_extras, _E_actual;
   double _missing;
 
 public:
   Manifold(){};
 
-  Manifold(std::unique_ptr<double[]>& flat, std::vector<double> y, size_t nobs, size_t E_x, size_t E_dt,
-           size_t E_extras, size_t E_actual, double missing)
+  Manifold(std::unique_ptr<double[]>& flat, std::vector<double> y, int nobs, int E_x, int E_dt, int E_extras,
+           int E_actual, double missing)
     : _flat(std::move(flat))
     , _y(y)
     , _nobs(nobs)
@@ -30,14 +30,14 @@ public:
     , _missing(missing)
   {}
 
-  double operator()(size_t i, size_t j) const { return _flat[i * _E_actual + j]; }
+  double operator()(int i, int j) const { return _flat[i * _E_actual + j]; }
 
-  double x(size_t i, size_t j) const { return _flat[i * _E_actual + j]; }
-  double dt(size_t i, size_t j) const { return _flat[i * _E_actual + _E_x + j]; }
-  double extras(size_t i, size_t j) const { return _flat[i * _E_actual + _E_x + _E_dt + j]; }
-  bool any_missing(size_t obsNum) const
+  double x(int i, int j) const { return _flat[i * _E_actual + j]; }
+  double dt(int i, int j) const { return _flat[i * _E_actual + _E_x + j]; }
+  double extras(int i, int j) const { return _flat[i * _E_actual + _E_x + _E_dt + j]; }
+  bool any_missing(int obsNum) const
   {
-    for (size_t j = 0; j < _E_actual; j++) {
+    for (int j = 0; j < _E_actual; j++) {
       if (operator()(obsNum, j) == _missing) {
         return true;
       }
@@ -45,9 +45,9 @@ public:
     return false;
   }
 
-  bool any_not_missing(size_t obsNum) const
+  bool any_not_missing(int obsNum) const
   {
-    for (size_t j = 0; j < _E_actual; j++) {
+    for (int j = 0; j < _E_actual; j++) {
       if (operator()(obsNum, j) != _missing) {
         return true;
       }
@@ -55,14 +55,14 @@ public:
     return false;
   }
 
-  double y(size_t i) const { return _y[i]; }
-  size_t ySize() const { return _y.size(); }
+  double y(int i) const { return _y[i]; }
+  int ySize() const { return (int)_y.size(); }
 
-  size_t nobs() const { return _nobs; }
-  size_t E() const { return _E_x; }
-  size_t E_dt() const { return _E_dt; }
-  size_t E_extra() const { return _E_extras; }
-  size_t E_actual() const { return _E_actual; }
+  int nobs() const { return _nobs; }
+  int E() const { return _E_x; }
+  int E_dt() const { return _E_dt; }
+  int E_extra() const { return _E_extras; }
+  int E_actual() const { return _E_actual; }
 };
 
 class ManifoldGenerator
@@ -73,11 +73,11 @@ private:
   bool _add_dt0 = false;
   int _tau;
   double _missing;
-  size_t _nobs;
-  size_t _num_extras, _num_extras_varying;
+  int _nobs;
+  int _num_extras, _num_extras_varying;
 
-  double lagged(const std::vector<double>& vec, const std::vector<size_t>& inds, size_t i, size_t j) const;
-  double find_dt(const std::vector<size_t>& inds, size_t i, size_t j) const;
+  double lagged(const std::vector<double>& vec, const std::vector<int>& inds, int i, int j) const;
+  double find_dt(const std::vector<int>& inds, int i, int j) const;
 
   double _dtWeight;
   std::vector<double> _x, _y, _co_x, _t;
@@ -94,7 +94,7 @@ public:
 
   ManifoldGenerator(const std::vector<double>& x, const std::vector<double>& y,
                     const std::vector<std::vector<double>>& extras, const std::vector<bool>& extrasEVarying,
-                    double missing, size_t tau)
+                    double missing, int tau)
     : _x(x)
     , _y(y)
     , _extras(extras)
@@ -102,7 +102,7 @@ public:
     , _missing(missing)
     , _tau(tau)
   {
-    _num_extras = extras.size();
+    _num_extras = (int)extras.size();
     _num_extras_varying = 0;
     for (const bool& eVarying : extrasEVarying) {
       _num_extras_varying += eVarying;
@@ -123,9 +123,9 @@ public:
     _add_dt0 = dt0;
   }
 
-  Manifold create_manifold(size_t E, const std::vector<bool>& filter, bool prediction) const;
+  Manifold create_manifold(int E, const std::vector<bool>& filter, bool prediction) const;
 
-  size_t E_dt(size_t E) const { return (_use_dt) * (E - 1 + _add_dt0); }
-  size_t E_extras(size_t E) const { return _num_extras + _num_extras_varying * (E - 1); }
-  size_t E_actual(size_t E) const { return E + E_dt(E) + E_extras(E); }
+  int E_dt(int E) const { return (_use_dt) * (E - 1 + _add_dt0); }
+  int E_extras(int E) const { return _num_extras + _num_extras_varying * (E - 1); }
+  int E_actual(int E) const { return E + E_dt(E) + E_extras(E); }
 };
